Feat: Create router management and change project structure
