+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "dart.flutterSdkPath": "/Users/mac/Documents/Software Development/SDK/flutter"
+}
