+++ lib/main.dart
@@ -1,7 +1,8 @@
-import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:splash_screen/ui/router/route_generator.dart';
+import 'package:splash_screen/ui/router/route_list.dart';
 
 void main() {
   runApp(MyApp());
@@ -11,65 +12,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: SplashScreen(),
+      title: "Flutter Splash",
+      initialRoute: routeSplashScreen,
+      onGenerateRoute: RouteGenerator.generate,
     );
   }
 }
 
-class SplashScreen extends StatefulWidget {
-  @override
-  _SplashScreenState createState() => _SplashScreenState();
-}
 
-class _SplashScreenState extends State<SplashScreen> {
-  @override
-  void initState() {
-    // TODO: implement initState
-    super.initState();
-    Timer(Duration(seconds: 3), () {
-      Navigator.of(context)
-          .pushReplacement(MaterialPageRoute(builder: (_) => HomePage()));
-    });
-  }
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xFF0B1328),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset(
-              'images/icon_trans.png',
-              height: 300.0,
-            ),
-            SizedBox(
-              height: 40.0,
-            ),
-            CircularProgressIndicator(
-              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Color(0xFF0B1328),
-        title: Text('Home Page'),
-      ),
-      body: Container(
-        child: Center(
-          child: Text('Home Page Content'),
-        ),
-      ),
-    );
-  }
-}
