+++ lib/ui/router/route_list.dart
@@ -1,0 +1,4 @@
+
+/// List of routing name
+const String routeSplashScreen = "/splashscreen";
+const String routeHomeScreen = "/home";
