+++ lib/ui/screens/splash/splash_screen.dart
@@ -1,0 +1,47 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:splash_screen/ui/router/route_list.dart';
+
+class SplashScreen extends StatefulWidget {
+  @override
+  _SplashScreenState createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+
+  void initialize() async {
+    await Future.delayed(Duration(seconds: 3));
+    Navigator.pushNamedAndRemoveUntil(context, routeHomeScreen, (route) => false);
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    this.initialize();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Color(0xFF0B1328),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Image.asset(
+              'images/icon_trans.png',
+              height: 300.0,
+            ),
+            SizedBox(
+              height: 40.0,
+            ),
+            CircularProgressIndicator(
+              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
